test(bench): ensure benchmarks fail on both a parsing error or empty yaml document
