+++ .gitignore
@@ -20,7 +20,7 @@
 !src/*.rs
 !build.rs
 !benches/*.rs
-!benches/fixtures/*.yaml
+!benches/fixtures/**/*.yaml
 target/**/*
 
 # Recurse through sub-directories applying the same patterns
