+++ Cargo.toml
@@ -25,5 +25,5 @@ built = { version = "0.8.0", features = ["git2", "semver", "chrono"] }
 criterion = { version = "0.7.0", features = ["html_reports"] }
 
 [[bench]]
-name = "parsing"
+name = "parser"
 harness = false
