+++ typos.toml
@@ -1,0 +1,4 @@
+[files]
+extend-exclude = [
+  "benches/fixtures/helm/**"
+]
